Fix Linux/GCC portability bugs found building fresh on Amazon Linux
None of these had ever been exercised by a from-scratch build before -
every prior build of this repo (including CI, presumably) reused an
existing macOS/clang build tree. Found while building vgi-sqlite (a
downstream client) fresh on Linux/aarch64 for the first time.

- arrow::util::base64_decode's return type is arrow::Result<std::string>
  on current vcpkg's Arrow (25.0.1), not the plain std::string this code
  assumed - catalog.cpp and function_dispatch.cpp both used the result
  directly without unwrapping. Fixed with .ValueOrDie(), matching this
  codebase's existing idiom for "trust this won't fail" Result handling
  (see catalog.cpp:71's arrow::MakeArrayOfNull(...).ValueOrDie()). Whether
  the older Arrow version this repo's cached macOS build linked against
  ever had a different (non-Result) signature, or this simply never
  compiled from scratch against a current Arrow, is unclear either way -
  the fix is correct for the API as it exists now.

- storage.cpp's highest_id() called std::max(int64_t, std::strtoll(...))
  - std::strtoll returns `long long`, which GCC's template argument
    deduction treats as distinct from `int64_t` (`long` on this LP64
    Linux target) even though both are 64-bit; clang apparently didn't
    balk at the same code. Fixed with an explicit static_cast<int64_t>.

- Two missing #include <stdexcept> (catalog.cpp, throwing
  std::invalid_argument/std::runtime_error; include/vgi/aggregate.h,
  throwing std::runtime_error) and one missing #include <algorithm>
  (arguments.cpp, using std::max/min) - all relied on transitively
  through some other header under libc++ (macOS's C++ standard library)
  but not under libstdc++ (GCC's, on Linux).

Scanned the rest of src/ and include/ for the same two bug classes
(base64_decode without unwrapping; std::max/min mixed with
strtoll/strtol) and for other files using stdexcept/algorithm symbols
without including them - no further instances found.

+++ src/storage.cpp
@@ -341,7 +341,7 @@ private:
         std::error_code ec;
         for (const auto& entry : fs::directory_iterator(dir, ec)) {
             const auto name = entry.path().filename().string();
-            highest = std::max(highest, std::strtoll(name.c_str(), nullptr, 10));
+            highest = std::max(highest, static_cast<int64_t>(std::strtoll(name.c_str(), nullptr, 10)));
         }
         return highest;
     }
